requirements.txt
-e git+https://github.com/fernandoe/django-fe-core.git@release/0.4.0#egg=django_fe_core
-e git+https://github.com/fernandoe/django-fe-jwt.git@release/0.0.1#egg=django_fe_jwt
dj-database-url==0.5.0
django-extensions==2.0.7
Django==2.0.5
djangorestframework-jwt==1.11.0
djangorestframework==3.8.2
factory-boy==2.10.0
mysqlclient==1.3.12
pytest-django==3.2.1
pytest==3.5.1
requests==2.18.4

## The following requirements were added by pip freeze:
attrs==17.4.0
certifi==2018.4.16
chardet==3.0.4
Faker==0.8.13
idna==2.6
more-itertools==4.1.0
pluggy==0.6.0
py==1.5.3
PyJWT==1.6.1
python-dateutil==2.7.2
pytz==2018.4
six==1.11.0
text-unidecode==1.2
urllib3==1.22
